+++ src/content-claims/api.ts
@@ -1,0 +1,3 @@
+export type Claim = import('@web3-storage/content-claims/client/api').Claim
+export type LocationClaim = import('@web3-storage/content-claims/client/api').LocationClaim
+export type RelationClaim = import('@web3-storage/content-claims/client/api').RelationClaim
